refactor(web): CaseList Code Splitting
